Introduce `lucet-module-data`.
This new crate provides a structured interface for manipulating, serializing, and deserializing
Lucet module data and metadata. It encodes this data using `bincode` via `serde`, with an emphasis
on using the zero-copy features of `serde` for potentially-large data like the initial heap. This
replaces code from both `lucet-runtime` and `lucetc`.

There is still more metadata to move into the purview of `lucet-module-data`, but this commit gets
us over the major humps of introducing the new package and adapting the test suites. In particular,
the interface for a mock module is changed substantially (for the better) in order to accommodate
the lifetime expectations of zero-copy data.

Closes #16.

Co-authored-by: Pat Hickey <phickey@fastly.com>