delete tests/!
these are now all duplicated in the lucet-runtime-tests, or test
something specific to the libc, which we'll replace in a test suite
specific to lucet-wasi.

the only one we dont have coverage for is the "session" suite, which
was always a hot mess and never found a bug. we mostly used that to
figure out how embeddings would work. we dont need it anymore.