[lucet-wasi] add `fd_fdstat_set_flags` and `path_open` syscalls
`path_open` in particular is the fundamental piece of the capability-based filesystem access. The
implementation in this patch is a pretty messy port of the CloudABI version; when we have more time
we should rewrite it in a more testable and Rusty fashion.

This patch also makes stdio handles inherited by default, though they can still be overridden with
other file descriptors. WASI assumes that file descriptors 0, 1, and 2 always exist.