[lucetc] tweak treatment of reserved size
Instead of a single `reserved-size` argument, `lucetc` now optionally takes `--min-reserved-size`
and `--max-reserved-size`. By default, the minimum is 4 KiB, and the maximum is 6 GiB (the entire
possible wasm address space, plus extra room for bounds check elision).

In the case that the module's initial heap size is larger than the minimum, we use the initial heap
size as the reserved size.

In the case that the initial size is larger than the maximum, compilation fails.