Build a "lucet" layer on top of "lucet-dev"
We end up with two images.

The former being just the base development environment, suitable for CI.

The later having the release builds populated in /opt/lucet.
This is the one we can upload to the Docker Hub, as people can use it
to get immediately started with Lucet without having to recompile it.

Also remove the wasi debian package after installation by the way.