lucet-runtime-tests: this crate only defines test functions for lucet-runtime
As a result, it includes a (and possibly more in the future) reference to
functions in lucet-runtime (currently, lucet_vmctx_get_heap), and cannot be
built independently of lucet-runtime. `cargo test` will attempt to do exactly
this and result in an unhelpful build error about undefined symbols, while
trying to produce a binary to test that is tightly coupled to lucet-runtime
anyway.

This change disables tests for lucet-runtime-tests to keep cargo from trying
to build tests for this crate.