Check that INSTANCE_PADDING is correct
This hard-coded padding size has been the cause of headaches in the
original C version, because it may need to be adjusted every time
a change is made to the `Instance` structure.

Furthermore, the context is platform-dependent, so the padding cannot
be one-size-fits-all.

The globals pointer must be right at the end of the page, and the
last member of the structure. This is documented, but not enforced in
code.

Reintroduce the checks making sure that this is the case.

The `INSTANCE_PADDING` constant remains fragile, and we can probably
find a better way to handle this. Even on Linux, I'm not convinced
that the current value is correct on non-x86_64 architectures.

The check can save quite a bit of debugging time when it's off.

Also add the current value for macOS.

We probably don't want a catchall value, as it is likely to be wrong
on untested platforms.