Replace static link flags with link options
Instead of hard-coded, linker-specific command-line flags, define a
portable set of options to express the intent.

These options are then converted into the correct set of command-line
flags for the given linker.

We may eventually want to do the same thing for the compiler, but this
is not really necessary for now since clang is all we support.