Find how a wasm32-wasi compatible LLVM and Clang were installed
The installation script now tries to compile a minimal program to
verify that a a compatible compilation toolchain exists.

We also try to deal with the fact that some environments require
a version suffix on clang and/or llvm tools (ex: Ubuntu 19.10),
but some don't (ex: macOS).

The script produces more verbose output about what it is doing
and what may be missing for a successful installation.