Improve import/export function support (#203)
* add (failing) test around exported imports of functions

* update function manifest entries with imported function addresses

this is performed by adding relocations to update function manifest pointers when the module is loaded, which itself this involves bumping `target_lexicon` to match what faerie and cranelift use.

this is necessary to support calling imported functions as an instance entrypoint

* add exported import test for lucet-runtime

this fails, but due to declaration issues with the artifact

* don't declare exported imports as export functions

exported imports will never have a corresponding function definition for
a function body local to the artifact. as a result, imported exports
cause a panic at artifact generation time

* add support for mock module builder to behave like an exported import is present

* add lucetc test around multiple import behavior

* rewrite write_relocated_slice to be clear about imported cases

* add some functions spectests depended on

this fixes the `func_ptrs` spectests!