pin `getrandom` to an early version to avoid pulling in `spin`
Newer versions of `getrandom` used a feature that caused `spin` to be used for synchronization by
`lazy_static`. This was causing a nondeterministic segfault in our signal handler tests.

This commit should be revisited once a new `getrandom` release is cut, since newer code no longer
pulls in this dependency.