[lucet-runtime] fix C resume API and address review comments
Since the yield/resume API for C does a sort of duck typing thing with null pointers, the new
dynamic typechecking was a little too strict. This makes the typechecking optional in a private
interface that the C API now uses.

In addition to some doc fixups, the factorial examples now return values normally in addition to
yielding intermediate values, hopefully reducing confusion about what the functions are doing.