[lucet-runtime] add `RunResult` and unify yield implementations
This ends the experiment with putting yields in `Err` so we could hang desperately onto `Result`,
but is worth the change.

This also introduces a zero-sized `EmptyYieldVal` to use as a placeholder when a yield/resume is not
providing/expecting a value. This lets us unify the implementation of the various yield operators.

Also changes the C yield/resume API to always provide and expect a wrapped pointer, simplifying the
yield implementations further.

Finally, includes some stray fixes to docs and exports, including a new method on `Module` so that
`lucet-wasi` doesn't have to import internals in its `main.rs`.