[lucet-runtime] refactor `State` and associated C APIs
These changes reduce the number of times that our types must be `Clone`able simply in order to keep
a copy around in the `State` while also returning a result. This had cascading implications, like
having to store termination and yield values in `Arc`s, which are all now gone. The only
restrictions on these values now is `Any + 'static`, which actually reflecs the semantics we care
about.

The downside to this is there are now more complicated invariants about states, but those are now
noted in docs on the struct.

Accommodating these changes in the C API meant ripping out the explicit `lucet_instance_state()`
function and replacing it with an out-parameter on run/resume. `lucet_state` is now `lucet_result`,
and is more analogous to `Result<RunResult, Error>`. Along the way, the pointers to strings for file
and symbol names were replaced with fixed-length buffers so that we don't need a free function for
`lucet_result`. This all makes the C API nicer, in addition to supporting the `State` refactoring.