[lucet-runtime] make State variants more explicit and document them
This applies the distributive property on the states involved in termination and yielding, so that
they no longer rely on having optional types with invariants we maintain manually.

This also moves the definition of `State` into its own module, along with docs explaining which
states can transition to other states.

At first, I tried adding a state transition function that would encode transition validity in
executable form, but adding the overhead of a function call turned out to be too much for an
operation that needs to be able to happen within a signal handler. Specifically, the compiler was
asking for a 77KiB stack probe upon entry to the transition function, which was blowing both guest
stacks and signal stacks.

I suspect something was up with the closure argument that was needed to let state transitions move
values from previous states.