Revisit termination model
Pretty happy with it now: reducing the number of flags being used to one
has made it very simple to decide when an instance is or is not able to
be safely terminated. `KillSwitch` still needs to grow return values to
indicate ability or inability to actually signal the guest.

This removes some variables and undoes a lot of test changes that were
made in the prior commit, as well.