the `guest_fault` tests in lucet-runtime fail for spooky signal reasons
it *seems* like signal handlers just don't stick sometimes. this is
influenced by the presence of --test-threads=1, in that if it is
present, or only one test is run, this does not fail. if n>1 tests are
present, and one of them is `sigsegv_handler_during_guest`, even if the
second test is empty (here, `hostcall_error`),
`sigsegv_handler_during_guest` is flaky and sometimes just doesn't
handle its segfault.