configure CI using Github Actions
there are two workflows:
* workflows/main.yml defines the Test and Lint jobs, which run in
  parallel on each push and PR.
    * Test takes about 19 minutes to complete.
    * Lint takes about 45 seconds to run `make indent-check`.
* workflows/fuzz.yml defines the Fuzz job, which runs nightly at
  midnight. It only runs on the master branch, so we need to merge this
  PR to test it. The fuzz smoke-test passes, so I expect that this will
  work. We will probably have to add some system to report failures it
  discovers. We may want to run it more often, as well.

there is one disabled workflow:
* workflows/mac-ci.yml.disabled will run the CI, except for fuzzing,
  under Mac OS. It is disabled because tests fail on Mac OS at the
  moment. It should be enabled as part of fixing Mac OS.

there is one Action:
* actions/test specifies (in action.yml) a job that builds the
  Dockerfile at the repo root and runs `make test` in the container.

The Test job uses the Dockerfile in the repo, but not the devenv scripts
to run it. Actions only runs Docker via the This means the devenv
scripts themselves have no CI under Actions.

Once we have Mac OS support enabled again, there is a lot less reason to
maintain the complex set of devenv scripts, which do a lot of fancy and
difficult-to-understand stuff to allow using `lucetc`, `lucet-wasi` and
other executables on the Mac (or windows, I guess).

We could instead document the one-liner invocations of Docker to
reproduce our standard developer environment locally.