do not reset terminable to true on all instance resumptions
Incorrectly had been queueing `lucet_context_activate` to execute on all
instance restarts, not just the initial execution of a function. This
resulted in setting `terminable` to true when resuming from a yielded
function, and would allow a second attempt to KillSwitch an instance.

Because this is only reachable while resuming from a hostcall, while the
instance had terminated during a yield in the hostcall, the worst that
could happen is that `terminable` would be true while terminating due to
a remote request to stop.

While it doesn't allow deadlocks or spurious signals, this state is
still unintended, so it is now fixed.