Timeouts (#150)
* add termination mechanism and use it to implement timeouts

The termination model is described in the `execution` module-level docs,
but to reiterate: when in guest code, terminate with a SIGALRM sent
immediately, and when in host code, terminate by setting a flag (setting
`execution_domain` to `Terminated`) and checking that flag when exiting
a hostcall.

Hostcalls currently have their safety, with respect to termination,
implemented by running the whole hostcall in `uninterruptable`, which
effectively makes all hostcalls critical sections. In the future, we may
want to expose the idea of interruptable hostcalls, both to avoid overhead
of synchronizing `KillState` when entering an exiting hostcalls, and to
avoid delayed termination.

In the extreme, even though this adds a tool for timeout-style logic,
it is at worst a request to stop execution, which requires hostcalls to
complete in a reasonable amount of time to guarantee that timeouts
are always witnessed promptly.