[lucet-runtime-internals] refactor guest-specific context data
This ended up changing around the interface to initializing and swapping contexts in some important
ways. Notably, a pointer to the parent context that `lucet_context_bootstrap` swaps into when the
entrypoint function returns is no longer written into the guest stack, but rather is a field on
`Context`. This field gets updated in the `to` context when calling `Context::swap()`, so that we
can swap `to` a context from different `from` contexts without trying to always swap back to the
original context.

This also gets us ready to abstract away instance/timeout-specific data from the `Context` struct by
making the use of a backstop callback function explicit but optional.