hotfix version of e252eae for the 0.4.1 branch; bump to 0.4.3
This doesn't add the interface for configuring signal stacks in order to maintain semver
compatibility, but fixes the bug and allows for the test suite to pass in debug mode.

For consistency, we're still updating all of the crate versions even though it was only
`lucet-runtime-internals` that changed for this release.

Version 0.4.2 existed for a time on the master branch before we made some further semver breaking
changes that meant we couldn't release it under that version. We're skipping it so that there's not
confusion between commits in that stretch of time, and this hotfix.