bump version to `0.5.2-dev`
This is a new thing we're trying in order to avoid phantom release numbers from appearing in the
repo history.

In the past, we would bump the version when landing a PR according to semver rules.  This led to a
situation where we had version `0.4.2` throughout the repo for some time without cutting a release,
before landing a bigger PR that required a change to `0.5.0`. So, when we needed to increase the
patch version for the `0.4.x` series, we ended up skipping `0.4.2` since it had existed for a time
elsewhere in the history.

The hope is that if a similar situation arises today, the only versions that would have existed in
the repo without being released are suffixed with `-dev`, avoiding this potential confusion.