Lucetc: fix macOS detection
Lucetc uses `target-lexicon` crate to detect the host OS.

On modern Macs, `target-lexicon` reports that the host is `Darwin`,
not `MacOSX`, at least I see this on my MacBook Pro running Catalina 10.15.2.

This leads to lucetc is not being able to set `LDFLAGS` correctly.

This change solves the issue by using the same LDFLAGS on both `Darwin` and `MacOSX`.