lucet-runtime: KillState does not need a terminable flag [..]
Because we are now viewing a pending instance as terminable, we use the
execution domain to indicate whether or not it is safe to issue a signal
to a running Instance.

This means that the `terminable` flag is no longer needed! We shouldn't
ever be in contention over this mutex, so this should not present a
notable performance hit.