lucet-runtime: `with_terminability` method [..]
This provides a few readability benefits in `run_func`. By outlining
this unsafe block into a helper method, we can attach comments about why
we are moving register values around before entry to the function.

Similarly, we move some comments from the assembly implementation of the
activation function, and into the `extern` block in
`lucet_runtime_internals::context`. This makes the implementation
details a little more discoverable for other developers, by moving
rationale for things like the choice of `rbx` out of the assembly and
into the Rust code.

The other benefit is that this lets us decouple the terminability
concerns from `lucet_context_activate`, which itself is *truly* only
concerned with running some sort of function upon entering a guest
context, not what the entry callback actually does.

This information can now live alongside the `Instance` methods that are
concerned with setting terminability before swapping into a guest
context.