move sigstack size enforcement to instance run, vs region create
This allows applications which want to manually control their signal stacks to use a value of `0`
for `limits.signal_stack_size` when creating their regions. If instances are set to automatically
install a signal stack are run with an insufficient size, it will fail at that point instead of when
the region is created.

The new test of manual sigstack management uses limits with zero sigstack size, to show that this
scenario works.

This commit adds a fair amount of documentation for these issues, though the big docstring in
`lib.rs` is starting to feel unwieldy.