🔧 limit compile-time table elements to the table's min
Key phrase from the [spec][spec]:

> The `min` size in the limits of the table type specifies the initial size of that table, while its
> `max`, if present, restricts the size to which it can grow later.

In other words, at compile time the largest a table is allowed to be is its `min` limit. The `max`
limit, if present, only limits the size to which the table can grow via `table.grow` instructions at
runtime.

Previously, if we found an element section that specified an index offset past the end of the table
vector, we would obligingly grow it longer, even if it exceeded the `min` limit of the table (and
possibly the `max` value as well?). This patch changes the behavior to allocate a vector of `min`
length and return an error if any element initializers would overflow it.

There's a new test in the `lucetc/tests/wasm` suite to check that we enforce this
limit. Distressingly, it does not appear that the spectests actually test this condition.

[spec]: https://webassembly.github.io/spec/core/syntax/modules.html#tables