use object instead of goblin in lucet-objdump (#474)
Over in #470[1], we discovered that `object` and `faerie` have different
approaches to writing relocations in shared objects.  In a nutshell,
`object`'s approach requires applying relocations to the shared object's
data whereas `faerie`'s does not (at least for the current structure of
`lucet-objdump`).

This change represents an intermediate step towards accommodating
`object`'s relocation approach by rewriting `lucet-objdump` to use
`object` instead of `goblin` for reading files.  `object` is a little
easier to work with, a little more abstract than `goblin`, and various
other bits of Lucet already use `object` anyway.

There are no differences in `lucet-objdump`'s output on a representative
file when `lucet-objdump` is compiled with and without this patch.

[1]: https://github.com/bytecodealliance/lucet/pull/470