in the guest_fault suite, check that signal handlers are restored
There was a bug in `illegal_instr_manuals_signal()` where the Lucet signal handler would remain
installed for the remainder of the test suite execution. Somehow this wasn't causing any failures,
but does if the test suite is run more than once.

To guard against this type of thing in the future, I added a wrapper to any tests that muck with the
process's signal handlers to ensure that the handlers are the same before and after the test runs.