lucetc: declare more things as local
From a functionality standpoint, this is not strictly necessary.
However, it does two things: it exports less from the static library,
which seems like a good thing, and it encourages `cranelift-object` to
more closely match `cranelift-faerie` in terms of the relocations it
uses in the final output, which is good for #470.