convert lucetc to use object as the I/O backend
This commit makes `lucetc` use `cranelift-object`/`object` as its
interface for writing out native objects.  The intent is that this moves
`lucetc` closer to getting PE/COFF support "for free".

We had discussed providing a runtime switch for compatibility/risk
reasons, but given that so much is going through `cranelift-module` now
and the amount of `object`-specific code is minimal, I believe that
providing a runtime switch is just extra complexity for no reason.
Happy to hear pushback on that.