do not reset kill state in signal handler
1. Overview

This patch changes where we re-initialize an instance's `KillState`
after completing an execution. This re-initialization, to review, is
required for upholding guarantees about the validity of a `KillSwitch`
with respect to an instance's lifetime.

2. Motivation

Previously, we reset the kill state in one of two places: (1) the lucet
context backstop `exit_guest_region`, and (2) the lucet signal handler,
`handle_signal` defined in
`lucet-runtime-internals/src/instance/signals.rs`.

The latter poses a potential problem, because neither `malloc` or `free`
are required to be async-signal-safe by POSIX.1 (see signal-safety(7)[1]
for more information on this topic).

This means that we should neither allocate or free a new kill state,
because `Arc` *does* in fact use heap allocations.[2]

3. Fix

Rather than resetting our kill state in two places, we handle this in
`swap_and_return` after jumping back to the host process. The kill state
is reset when handling our `State` transition, *unless* the guest is
yielded.

Outstanding kill switches must not be invalidated when a guest yields.
In order to accomodate nicer control flow in `swap_and_return`, the
`State` enum has grown a new `is_yielding` method.

[1]: http://man7.org/linux/man-pages/man7/signal-safety.7.html
[2]: https://doc.rust-lang.org/src/alloc/sync.rs.html#299-308

Co-authored-by: iximeow git@iximeow.net