🎬 Add explicit API to run start functions; don't run it implicitly
Previously, the start function would be implicitly run upon instance creation or resetting. Since
the environment can be different at instantiation time to when export functions are typically
called (missing embedder contexts, etc), this has tripped people up. It can also just be surprising
that any code runs at all at those points, which combined with the inherently confusing nature of
Wasm start functions just led to too many messes.

`Instance::run_start()` is now part of the public API, and will run the start function if it is
present in that instance's Wasm module. It does nothing if there is no start
function. `Instance::run()` will now return `Err(Error::InstanceNeedsStart)` if the start function
is present but hasn't been run since the instance was created or reset.

This commit bumps the development version to `0.7.0-dev`, because while the API changes are
semver-compatible, the semantics are not.