move mmap into sysdeps
The patch is relatively easy; what's not clear to me is how something
like this really wants to be exported.  Should `sysdeps` export a
`MappedRegion` that would be backed by `VirtualAlloc` on Windows?
Should `sysdeps` just export everything with their platform names and
let `lucet-runtime` use conditionals to choose what to export?

This patch punts on those questions for now; we're just trying to move
the obvious system-dependent code into `sysdeps`.