🎬 Encode the start function in ModuleData rather than the linker
This long-overdue refactoring moves the metadata required to get a module's Wasm start function to
the `lucet-module` types, rather than encoding it directly as a pointer-to-a-pointer in the shared
object's symbols.

Of particular benefit is the fact that once a `DlModule` is created, we no longer use the underlying
`libloading::Library`. Before the version bump in this patch, that library had a mutex around calls
to `dlerror` to try preventing races between all threads in the process, which could lead to
contention when threads attempted to concurrently use Lucet instances.