Remove AssemblyScript examples
Examples of using AssemblyScript with Lucet was quite nice to have
for Terrarium.

But AssemblyScript has changed a lot since, it's got great support
for WASI, `as-wasi` is now available on npm, and there's no point
in keeping copies of these things in the Lucet tree any more.

If people specifically ask about how to use AssemblyScript with
Lucet, relevant documentation can still be added to these projects
own documentation.