lucetc: cranelift 0.64 tracks signatures differently.
this is actually a really nice improvement to signatures: rather than
having to reverse-engineer a wasm func type from an ir::Signature,
we get access to the actual FuncType right from the wasm module. So,
keep that around in lucetc's ModuleInfo, and that vastly simplifies
creating the lucet-module signature as well.

Renamed some error types for clarity, as well.