💢 Package unknown hostcall panics into termination details
This prevents panics that aren't intentionally initiated by the Lucet runtime from attempting to
unwind across Wasm stack frames. When #254 lands, this will no longer be necessary, but until then
this allows us to safely transport the unknown panic payload across the FFI boundary so that
panicking can resume on the other side.